requirements.txt
crewai
crewai_tools
langchain_community
langchain_openai
python-dotenv
duckduckgo-search
arxiv
tavily-python
langchain-google-genai
streamlit
pysqlite3-binary